use middleware to handle poll sync
