+++ react/features/poll/components/PollPanel/web/PollPanel.js
@@ -5,8 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { ThemeProvider } from 'styled-components';
 
-import { SYNC_POLL } from '../../../actionTypes';
-import { close, setSyncToFalse } from '../../../actions';
+import { close } from '../../../actions';
 import { classList, getPollsPaneOpen } from '../../../functions';
 import theme from '../../../theme.json';
 import { PollsList, PollCreation, PollDetail } from '../../PollsList/';
@@ -24,7 +23,6 @@ const PollsPane = () => {
     const { t } = useTranslation();
 
     const closePane = useCallback(() => dispatch(close(), [ dispatch ]));
-    const turnSyncOff = useCallback(() => dispatch(setSyncToFalse(), [ dispatch ]));
     const pollPaneMode = useSelector(state => state['features/poll'].pollPaneMode);
 
     const closePaneKeyPress = useCallback(e => {
@@ -33,30 +31,6 @@ const PollsPane = () => {
             closePane();
         }
     }, [ closePane ]);
-
-    const sendSyncMsg = useSelector(state => state['features/poll'].sendSyncMsg);
-    const conference = useSelector(state => state['features/base/conference'].conference);
-
-
-    const sendSyncMessage = useCallback((polls, pollResponses, optionsList) => {
-
-        const msg = { polls,
-            pollResponses,
-            optionsList,
-            type: SYNC_POLL };
-
-        if (conference) {
-            conference.sendMessage(msg);
-        }
-
-    });
-
-    const { polls, pollResponses, optionsList } = useSelector(state => state['features/poll']);
-
-    if (sendSyncMsg) {
-        sendSyncMessage(polls, pollResponses, optionsList);
-        turnSyncOff();
-    }
 
     const renderContent = () => {
         switch (pollPaneMode) {
